Make shared KubeadmConfig fields reusable
In preparation for introducing provisioner-flavored CABPK
implementations, we want to reuse common fields in KubeadmConfigSpec.

KubeadmConfigSpec contains two categories of fields:

1. Fields such as ClusterConfiguration which contain generic data
   related to cluster bootstrapping.
2. Fields such as DiskSetup which contain cloud-init-specific data.

If we are to add e.g. an Ignition-flavored CABPK, we want to reuse the
generic fields while replacing cloud-init fields with Ignition fields.

This change should be transparent to users because through the use of Go
embedding we modify the structs underlying the KubeadmConfig CRD without
changing the API layer (note that we aren't modifying anything under
config/crd).

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>